updated gov pay session validation
